+++ qbraid/interface/programs.py
@@ -35,7 +35,8 @@ def bell_data() -> QROGRAM_TEST_TYPE:
     from qbraid.interface.qbraid_cirq.circuits import cirq_bell
     from qbraid.interface.qbraid_pyquil.programs import pyquil_bell
     from qbraid.interface.qbraid_pytket.circuits import pytket_bell
-    from qbraid.interface.qbraid_qasm.circuits import qasm2_bell, qasm3_bell
+    from qbraid.interface.qbraid_qasm3.circuits import qasm3_bell
+    from qbraid.interface.qbraid_qasm.circuits import qasm2_bell
     from qbraid.interface.qbraid_qiskit.circuits import qiskit_bell
 
     unitary = to_unitary(cirq_bell())
@@ -58,7 +59,8 @@ def shared15_data() -> QROGRAM_TEST_TYPE:
     from qbraid.interface.qbraid_braket.circuits import braket_shared15
     from qbraid.interface.qbraid_cirq.circuits import cirq_shared15
     from qbraid.interface.qbraid_pytket.circuits import pytket_shared15
-    from qbraid.interface.qbraid_qasm.circuits import qasm2_raw_shared15, qasm3_shared15
+    from qbraid.interface.qbraid_qasm3.circuits import qasm3_shared15
+    from qbraid.interface.qbraid_qasm.circuits import qasm2_raw_shared15
     from qbraid.interface.qbraid_qiskit.circuits import qiskit_shared15
 
     unitary = to_unitary(cirq_shared15())
@@ -100,7 +102,7 @@ def random_circuit(
     num_qubits = np.random.randint(1, 4) if num_qubits is None else num_qubits
     depth = np.random.randint(1, 4) if depth is None else depth
     if package == "qasm3":
-        from qbraid.interface.qbraid_qasm.circuits import _qasm3_random
+        from qbraid.interface.qbraid_qasm3.circuits import _qasm3_random
 
         rand_circuit = _qasm3_random(num_qubits, depth, **kwargs)
     elif package == "qiskit":
