+++ qbraid/interface/qbraid_qasm/tools.py
@@ -9,10 +9,9 @@
 # THERE IS NO WARRANTY for the qBraid-SDK, as per Section 15 of the GPL v3.
 
 """
-Module containing OpenQasm tools
+Module containing OpenQASM 2 tools
 
 """
-import os
 import re
 from collections import defaultdict
 from typing import List
@@ -20,7 +19,6 @@ from typing import List
 import numpy as np
 
 from qbraid.interface.qbraid_qasm.qasm_preprocess import convert_to_supported_qasm
-from qbraid.interface.qbraid_qasm.qelib1_defs import _decompose_rxx_instr
 
 QASMType = str
 
@@ -29,14 +27,14 @@ def qasm_qubits(qasmstr: str) -> List[QASMType]:
     """Get number of qasm qubits.
 
     Args:
-        qasmstr (str): OpenQASM 2 or OpenQASM 3 string
+        qasmstr (str): OpenQASM 2 string
 
     Returns:
         List of qubits in the circuit
     """
     return [
         text.replace("\n", "")
-        for match in re.findall(r"(\bqreg\s\S+\s+\b)|(qubit\[(\d+)\])|(\bqubit\b)", qasmstr)
+        for match in re.findall(r"(\bqreg\s\S+\s+\b)|(qubit\[(\d+)\])", qasmstr)
         for text in match
         if text != "" and len(text) >= 2
     ]
@@ -54,23 +52,6 @@ def qasm_num_qubits(qasmstr: str) -> int:
     return q_num
 
 
-def qasm_3_num_qubits(qasmstr: str) -> int:
-    """Calculate number of qubits in a qasm3 string"""
-    q_num = 0
-    for bit_line in qasm_qubits(qasmstr):
-        if bit_line != "qubit":
-            # multiple qubits
-
-            # split is needed as the name may contain
-            # a number
-            bit_line = bit_line.split("[")[1]
-            q_num += int(re.search(r"\d+", bit_line).group())
-        else:
-            # single qubit
-            q_num += 1
-    return q_num
-
-
 def qasm_depth(qasmstr: str) -> int:
     """Calculates circuit depth of OpenQASM 2 string"""
     qasm_input = convert_to_supported_qasm(qasmstr)
@@ -79,7 +60,8 @@ def qasm_depth(qasmstr: str) -> int:
     gate_lines = [
         s
         for s in lines
-        if s.strip() and not s.startswith(("OPENQASM", "include", "qreg", "gate", "//"))
+        if s.strip()
+        and not s.startswith(("OPENQASM", "include", "qreg", "gate", "measure", "if", "//"))
     ]
 
     counts_dict = defaultdict(int)
@@ -95,14 +77,6 @@ def qasm_depth(qasmstr: str) -> int:
             counts_dict[f"q[{i}]"] = max_depth + 1
 
     return max(counts_dict.values()) if counts_dict else 0
-
-
-def qasm3_depth(qasmstr: str) -> int:
-    """Calculates circuit depth of OpenQASM 3 string"""
-    # pylint: disable=import-outside-toplevel
-    from qiskit.qasm3 import loads
-
-    return loads(qasmstr).depth()
 
 
 def _convert_to_contiguous_qasm(qasmstr: str, rev_qubits=False) -> QASMType:
@@ -121,106 +95,3 @@ def _unitary_from_qasm(qasmstr: QASMType) -> np.ndarray:
     from qbraid.transpiler.cirq_qasm.qasm_conversions import from_qasm
 
     return from_qasm(qasmstr).unitary()
-
-
-def _build_qasm_3_reg(line: str, qreg_type: bool) -> QASMType:
-    """Helper function to build openqasm 3 register statements
-
-    Args:
-        line (str): openqasm 2  regdecl statement
-        qreg_type (bool): whether a qreg or creg type statement
-
-    Returns:
-        str : openqasm 3 qubits / bits declaration
-    """
-    reg_keyword_len = 4
-    line = line[reg_keyword_len:]
-    elements = line.split("[")
-    reg_name = elements[0].strip()
-    reg_size = elements[1].split("]")[0].strip()
-    result = "qubit" if qreg_type else "bit"
-    result += f"[{reg_size}] {reg_name};\n"
-    return result
-
-
-def _build_qasm_3_measure(line: str) -> QASMType:
-    """Helper function to build openqasm 3 measure string
-
-    Args:
-        line (str): openqasm 2 measure statement
-
-    Returns:
-        str:  openqasm 3 measure statement
-    """
-    measure_keyword_len = 7
-    line = line[measure_keyword_len:]
-    elements = line.split("->")
-    qubits_name = elements[0].replace(" ", "")
-    bits_name = elements[1].split(";")[0].replace(" ", "")
-
-    return f"{bits_name} = measure {qubits_name};\n"
-
-
-def _change_to_qasm_3(line: str) -> QASMType:
-    """Function to change an openqasm 2 line to openqasm 3
-
-    Args:
-        line (str): an openqasm 2 line
-
-    Returns:
-        str: corresponding openqasm 3 line
-    """
-    line = line.lstrip()
-    if line.startswith("OPENQASM"):
-        return ""
-    if "qelib1.inc" in line:
-        return ""
-    if line.startswith("qreg"):
-        return _build_qasm_3_reg(line, qreg_type=True)
-    if line.startswith("creg"):
-        return _build_qasm_3_reg(line, qreg_type=False)
-    if line.startswith("u("):
-        return line.replace("u(", "U(")
-    if line.startswith("rxx("):
-        return _decompose_rxx_instr(line)
-    if line.startswith("measure"):
-        return _build_qasm_3_measure(line)
-    if line.startswith("opaque"):
-        # as opaque is ignored by openqasm 3 add it as a comment
-        return "// " + line + "\n"
-    return line + "\n"
-
-
-def convert_to_qasm3(qasm2_str: str):
-    """Convert a QASM 2.0 string to QASM 3.0 string
-
-    Args:
-        qasm2_str (str): QASM 2.0 string
-    """
-    # pylint: disable=import-outside-toplevel
-    from qiskit import QuantumCircuit
-
-    try:
-        # use inbuilt method to check validity
-        _ = QuantumCircuit.from_qasm_str(qasm2_str)
-    except Exception as e:
-        raise ValueError("Invalid QASM 2.0 string") from e
-
-    #  a newline separated qasm 2 string
-    # formatted_qasm2 = circuit.qasm()
-    qasm3_str = """OPENQASM 3.0;
-include 'stdgates.inc';"""
-
-    # add the gate from qelib1.inc not present in the
-    # stdgates.inc file
-    current_dir = os.path.dirname(os.path.abspath(__file__))
-    with open(
-        os.path.join(current_dir, "qasm_lib/qelib_qasm3.qasm"), mode="r", encoding="utf-8"
-    ) as gate_defs:
-        for line in gate_defs:
-            qasm3_str += line
-
-    for line in qasm2_str.splitlines():
-        line = _change_to_qasm_3(line)
-        qasm3_str += line
-    return qasm3_str
